requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@ca4db6868655dfe6499807121651f4e5a7f5fa6e
frozenlist==1.3.0
idna==3.3
mysql-connector-python==8.0.28
multidict==6.0.2
python-decouple==3.6
yarl==1.7.2